Reorganization of the codebase
I reorganized JSX files into pages and components. Also changed About
page again to actually fit the mobile screen.
