+++ frontend/src/App.jsx
@@ -1,12 +1,12 @@
-import Register from './components/Register';
-import About from './components/About';
-import Login from './components/Login';
-import Verify from './components/Verify'
+import Register from './pages/Register';
+import About from './pages/About';
+import Login from './pages/Login';
+import Verify from './pages/Verify'
 import NavBar from './components/Navbar'
-import Forgot from './components/Forgot'
-import Recover from './components/Recover'
-import SpotifyStats from './components/SpotifyStats'
-import Discover from './components/Discover'
+import Forgot from './pages/Forgot'
+import Recover from './pages/Recover'
+import SpotifyStats from './pages/SpotifyStats'
+import Discover from './pages/Discover'
 import LandingPage from './pages/Landing'
 import ProfilePage from './pages/Profile';
 import ErrorBar from './components/ErrorBar';
