+++ frontend/src/pages/Discover.jsx
