+++ frontend/src/pages/SpotifyStats.jsx
